Remakes the websocket service to be fully stateless.
The websocket service is now fully stateless. In a scaled enironment
each instance of the websocket-service has it's own kafka consumer id,
so each instance will now receive a published message for the webscoket,
but only the instance, which has the websocket connection localy will
process and send the message to the user, the other instances will discard it.
The frontend now automaticly reconnects to an available websocket service session, when
its own session was terminated due to service downscaleing.